wr-193: * rename SharedOnFacebook button folder name
